SWATCH-2296: Remove the "swatch-subscription-sync" clowadpp
