+++ api/pom.xml
@@ -16,8 +16,6 @@
   <properties>
     <api_spec_v1>${project.basedir}/rhsm-subscriptions-api-v1-spec.yaml</api_spec_v1>
     <api_config_v1>${project.basedir}/rhsm-subscriptions-api-v1-config.json</api_config_v1>
-    <api_spec_v2>${project.basedir}/rhsm-subscriptions-api-v2-spec.yaml</api_spec_v2>
-    <api_config_v2>${project.basedir}/rhsm-subscriptions-api-v2-config.json</api_config_v2>
   </properties>
 
   <dependencies>
@@ -118,68 +116,6 @@
               </configOptions>
             </configuration>
           </execution>
-
-          <execution>
-            <id>generateApi_v2</id>
-            <goals>
-              <goal>generate</goal>
-            </goals>
-            <configuration>
-              <inputSpec>${api_spec_v2}</inputSpec>
-              <generatorName>jaxrs-spec</generatorName>
-              <configurationFile>${api_config_v2}</configurationFile>
-              <configOptions>
-                <sourceFolder>src/gen/java/main</sourceFolder>
-                <interfaceOnly>true</interfaceOnly>
-                <generatePom>false</generatePom>
-                <dateLibrary>java8</dateLibrary>
-                <useJakartaEe>true</useJakartaEe>
-                <useTags>true</useTags>
-              </configOptions>
-              <typeMappings>
-                <typeMapping>string+MetricId=MetricId</typeMapping>
-                <typeMapping>string+ProductId=ProductId</typeMapping>
-              </typeMappings>
-              <importMappings>
-                <importMapping>StreamingOutput=javax.ws.rs.core.StreamingOutput</importMapping>
-                <importMapping>MetricId=com.redhat.swatch.configuration.registry.MetricId
-                </importMapping>
-                <importMapping>ProductId=com.redhat.swatch.configuration.registry.ProductId
-                </importMapping>
-              </importMappings>
-            </configuration>
-          </execution>
-
-          <execution>
-            <id>generateApiDocs_v2</id>
-            <goals>
-              <goal>generate</goal>
-            </goals>
-            <configuration>
-              <generatorName>html</generatorName>
-              <inputSpec>${api_spec_v2}</inputSpec>
-              <output>${project.build.directory}/docs</output>
-              <generateModelTests>false</generateModelTests>
-              <generateApiTests>false</generateApiTests>
-            </configuration>
-          </execution>
-
-          <execution>
-            <id>generateOpenApiJson_v2</id>
-            <goals>
-              <goal>generate</goal>
-            </goals>
-            <configuration>
-              <generatorName>openapi</generatorName>
-              <inputSpec>${api_spec_v2}</inputSpec>
-              <output>${project.build.directory}/openapijson</output>
-              <generateModelTests>false</generateModelTests>
-              <generateApiTests>false</generateApiTests>
-              <configOptions>
-                <outputFileName>rhsm-subscriptions-api-v2-openapi.json</outputFileName>
-              </configOptions>
-            </configuration>
-          </execution>
         </executions>
       </plugin>
 
@@ -212,34 +148,6 @@
               <output>${project.build.directory}/classes/swagger</output>
               <configOptions>
                 <outputFile>generated-api-spec-v1.json</outputFile>
-              </configOptions>
-            </configuration>
-          </execution>
-          <execution>
-            <id>generate-api-v2-spec-yaml</id>
-            <goals>
-              <goal>generate</goal>
-            </goals>
-            <configuration>
-              <inputSpec>${api_spec_v2}</inputSpec>
-              <language>openapi-yaml</language>
-              <output>${project.build.directory}/classes/swagger</output>
-              <configOptions>
-                <outputFile>generated-api-spec-v2.yaml</outputFile>
-              </configOptions>
-            </configuration>
-          </execution>
-          <execution>
-            <id>generate-api-v2-spec-json</id>
-            <goals>
-              <goal>generate</goal>
-            </goals>
-            <configuration>
-              <inputSpec>${api_spec_v2}</inputSpec>
-              <language>openapi</language>
-              <output>${project.build.directory}/classes/swagger</output>
-              <configOptions>
-                <outputFile>generated-api-spec-v2.json</outputFile>
               </configOptions>
             </configuration>
           </execution>
