+++ api/rhsm-subscriptions-api-v1-spec.yaml
@@ -588,161 +588,6 @@ paths:
           $ref: "../spec/error-responses.yaml#/$defs/InternalServerError"
       tags:
         - contracts
-  /v1/subscriptions/products/{product_id}:
-    description: "Operations for total capacity by SKU for all of the org's active subscriptions for given Swatch product ID."
-    parameters:
-      - name: product_id
-        in: path
-        required: true
-        schema:
-          $ref: "#/components/schemas/ProductId"
-        description: "The ID for the product we wish to query"
-      - name: offset
-        in: query
-        schema:
-          type: integer
-          minimum: 0
-        description: "The number of items to skip before starting to collect the result set"
-      - name: limit
-        in: query
-        schema:
-          type: integer
-          minimum: 1
-        description: "The numbers of items to return"
-    get:
-      summary: "Returns the total capacity by SKU for all of the org's active subscriptions for given Swatch product ID."
-      operationId: getSkuCapacityReport
-      deprecated: true
-      parameters:
-        - name: category
-          in: query
-          schema:
-            $ref: '#/components/schemas/ReportCategory'
-          description: 'The category to fetch data for'
-        - name: sla
-          in: query
-          schema:
-            $ref: '#/components/schemas/ServiceLevelType'
-          description: "Include only capacity for the specified service level."
-        - name: usage
-          in: query
-          schema:
-            $ref: '#/components/schemas/UsageType'
-          description: "Include only capacity for the specified usage level."
-        - name: billing_provider
-          in: query
-          schema:
-            $ref: '#/components/schemas/BillingProviderType'
-          description: "Include only report data matching the specified billing provider"
-        - name: billing_account_id
-          in: query
-          schema:
-            $ref: '#/components/schemas/BillingAccountId'
-          description: "Include only report data matching the specified billing account"
-        - name: beginning
-          in: query
-          schema:
-            type: string
-            format: date-time
-          description: "Defines the start of the report period. Dates should be provided in ISO 8601
-             format but the only accepted offset is UTC. E.g. 2017-07-21T17:32:28Z"
-        - name: ending
-          in: query
-          schema:
-            type: string
-            format: date-time
-          description: "Defines the end of the report period.  Defaults to the current time. Dates should
-             be provided in UTC."
-        - name: metric_id
-          in: query
-          schema:
-            type: string
-          description: "Filter subscriptions to those that contribute to a specific unit of measure"
-        - name: sort
-          in: query
-          schema:
-            $ref: "#/components/schemas/SkuCapacityReportSort"
-          description: "What property to sort by (optional)"
-        - name: dir
-          in: query
-          schema:
-            $ref: "#/components/schemas/SortDirection"
-          description: "Which direction to sort by (default: asc)"
-      responses:
-        '200':
-          description: "The request for the account's subscription capacities was successful."
-          content:
-            application/vnd.api+json:
-              schema:
-                $ref: "#/components/schemas/SkuCapacityReport"
-              example:
-                data:
-                  - sku: RH00011
-                    product_name: "Red Hat Enterprise Linux Server, Premium (Physical and 4 Virtual Nodes)(L3 Only)"
-                    service_level: Premium
-                    usage: Production
-                    subscriptions:
-                      - id: "1234567890"
-                        number: "1234567891"
-                      - id: "1234567892"
-                        number: "1234567893"
-                      - id: "1234567894"
-                        number: "1234567895"
-                    next_event_date: "2020-04-01T00:00:00Z"
-                    next_event_type: Subscription Begin
-                    quantity: 3
-                    capacity: 3
-                    hypervisor_capacity: 3
-                    total_capacity: 6
-                    metricId: Sockets
-                  - sku: RH00010
-                    product_name: "Red Hat Enterprise Linux Server"
-                    service_level: Self-Support
-                    usage: Production
-                    subscriptions:
-                      - id: "1234567896"
-                        number: "1234567897"
-                      - id: "1234567898"
-                        number: "1234567899"
-                      - id: "1234567900"
-                        number: "1234567901"
-                    next_event_date: "2020-04-02T00:00:00Z"
-                    next_event_type: Subscription Begin
-                    quantity: 3
-                    capacity: 3
-                    hypervisor_capacity: 3
-                    total_capacity": 6
-                    metricId: Sockets
-                  - sku: RH00009
-                    product_name: "Red Hat Enterprise Linux Server, Premium"
-                    service_level: Premium
-                    usage: Production
-                    subscriptions:
-                      - id: "1234567902"
-                        number: "1234567903"
-                      - id: "1234567904"
-                        number: "1234567905"
-                      - id: "1234567906"
-                        number: "1234567907"
-                    next_event_date: "2020-04-01T00:00:00Z"
-                    next_event_type: Subscription End
-                    quantity: 3
-                    capacity: 6
-                    hypervisor_capacity: 6
-                    total_capacity: 12
-                    metricId: Cores
-                meta:
-                  count: 3
-        '400':
-          $ref: "../spec/error-responses.yaml#/$defs/BadRequest"
-        '403':
-          $ref: "../spec/error-responses.yaml#/$defs/Forbidden"
-        '404':
-          $ref: "../spec/error-responses.yaml#/$defs/ResourceNotFound"
-        '500':
-          $ref: "../spec/error-responses.yaml#/$defs/InternalServerError"
-      tags:
-        - subscriptions
   /v1/openapi.json:
     $ref: "../spec/openapi-paths.yaml#/openapi-json"
   /v1/openapi.yaml:
