--- src/main/resources/application-liquibase-only.yaml
@@ -1,1 +1,0 @@
-POSTGRES_STATEMENT_TIMEOUT: 0
