+++ src/main/resources/application.yaml
@@ -10,8 +10,6 @@ spring:
   profiles:
     active:
       - api
-      - capacity-ingress
-      - rhsm-conduit
       - worker
       - kafka-queue
   jms:
