+++ src/test/java/org/candlepin/subscriptions/resource/api/SwaggerTest.java
@@ -36,7 +36,7 @@ import org.springframework.test.context.ActiveProfiles;
  * that point out to the correct openapi endpoints).
  */
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("test")
+@ActiveProfiles({"api", "test"})
 class SwaggerTest {
 
   private static final String LOCALHOST = "http://localhost:";
