+++ src/test/java/org/candlepin/subscriptions/tally/TallySnapshotControllerIT.java
@@ -78,7 +78,7 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest(properties = {"HOURLY_TALLY_EVENT_BATCH_SIZE=2"})
-@ActiveProfiles(value = {"worker", "kafka-queue", "api", "test-inventory", "capacity-ingress"})
+@ActiveProfiles(value = {"worker", "kafka-queue", "api", "test-inventory"})
 @ExtendWith(OutputCaptureExtension.class)
 class TallySnapshotControllerIT implements ExtendWithSwatchDatabase, ExtendWithEmbeddedKafka {
   static final String INSTANCE_ID = "i-123456";
