+++ swatch-core-test/src/main/resources/application-test.yaml
@@ -12,7 +12,6 @@ rhsm-subscriptions:
     driver-class-name: org.hsqldb.jdbc.JDBCDriver
     platform: hsqldb
   account-batch-size: 2
-  subscription-sync-enabled: true
 logging:
   level:
     org:
